+++ src/components/drawer.ts
@@ -1,0 +1,2 @@
+export * from './drawer/drawer.component.js';
+export { default } from './drawer/drawer.component.js';